+++ src/main/kotlin/com/app/drivn/backend/user/dto/UserInfoDto.kt
@@ -5,9 +5,9 @@ import java.math.BigDecimal
 import java.time.ZonedDateTime
 
 data class UserInfoDto(
-  val distance: Float,
-  val energy: Float,
-  val maxEnergy: Float,
+  val distance: BigDecimal,
+  val energy: BigDecimal,
+  val maxEnergy: BigDecimal,
   val tokenClaimable: BigDecimal,
   val tokensLimitPerDay: BigDecimal,
   val nextEnergyRenew: ZonedDateTime?,
